PHEC (Dec 2016) Generalised Clipped Boundaries in England

This data contains the digital vector boundaries for Public Health England Centres in England as at December 2016. The boundaries are generalised (20m) - clipped to the coastline (Mean High Water mark). Contains both Ordnance Survey and ONS Intellectual Property Rights.

Usage

PHEC_boundaries_2016

Format

A dataset with variables:

OBJECTID

phec16cd

Public Health England Centre code

phec16nm

Public Health England Centre name

bng_e

bng_n

long

lat

GlobalID

SHAPE_Length

SHAPE_Area

geometry

Source

Source: Office for National Statistics licensed under the Open Government Licence v.3.0 Contains OS data © Crown copyright and database right 2022 https://open-geography-portalx-ons.hub.arcgis.com/datasets/ons::phec-dec-2016-generalised-clipped-boundaries-in-england/explore