Age-Sex Pyramid using Line List Data

This function creates an age-sex pyramid using line list data.

agesex_pyramid_linelist(
  df,
  age_breakpoints = c(0, 5, 19, 65, Inf),
  colours = c("#440154", "#fde725"),
  x_breaks = 10,
  y_title = "Proportion of People Percentage (%)",
  text_size = 12,
  conf_limits = FALSE,
  age_calc_refdate = Sys.Date()
)

Arguments

df

A data frame containing line list data.

age_breakpoints

A numeric vector specifying the age breakpoints for defining age groups. Default is c(0, 4, 18, 64, Inf).

colours

A character vector specifying the colors for the genders in HEX format, in the order Female, Male. Default is c("#440154", "#fde725").

x_breaks

Number of ticks on the X-axis. Default is 20.

y_title

Title for the Y-axis. Default is "Proportion of People Percentage (%)".

text_size

Text size for the plot. Default is 12.

conf_limits

Logical, indicating whether to include confidence limits (error bars) in the plot. Default is FALSE.

age_calc_refdate

Reference date for calculating age from date of birth if 'age' column is not available in the dataset. Default is Sys.Date() (today's date).

Value

Returns an age-sex pyramid as a ggplot object.

Examples

if (FALSE) {
  # Sample line list data
  data <- data.frame(
    date_of_birth = as.Date(c("1990-05-15", "1985-08-22", "1972-03-10", "2002-11-30")),
    sex = c("Male", "Female", "Male", "Male")
  )

  # Generate age-sex pyramid
  agesex_pyramid_linelist(data)
}