Line_chart

The chart function has 3 mandatory arguments to plot, with additional arguments as described further below:

line_chart_plotly(
  df,
  x,
  y,
  grouping_col,
  plot_mode = "lines",
  color_theme = c("skyblue", "lightgreen", "lightcoral", "grey", "gold", "plum", "tan"),
  y_label = "",
  yaxis_label_angle = 0,
  y_rangemode = "tozero",
  y_min_limit = NULL,
  y_max_limit = NULL,
  y_as_percent = FALSE,
  x_label = "",
  x_axis_reverse = FALSE,
  xaxis_label_angle = 90,
  lower_limit = NULL,
  upper_limit = NULL,
  hover_text = NULL,
  legend_orientation = "h",
  legend_x = 0.5,
  legend_y = -0.35,
  show_legend = TRUE
)

Arguments

df

name of the dataframe plotting, and are mandatory.

x

names of the x axis variable of the plot from the df and are mandatory.

y

names of the y axis variable of the plot from the df and are mandatory.

grouping_col

names of the grouping_col variable of the plot from the df

plot_mode

default is "line" use "line+marker" if needed

color_theme

is array of colours

y_label

y axis title default is nothing , fill in needed

yaxis_label_angle

default is 0, can set to other eg:-45,

y_rangemode

set to "tozero",

y_min_limit

set to the min of the y value, default is null

y_max_limit

set to the max of the y value, default is null

y_as_percent

default is FALSE, Only TRUE when need y as percentage

x_label

xaxis title default is "", add the parameter if needed

x_axis_reverse

default is False, only TRUE when you want to reverse the xaxis

xaxis_label_angle

xaxis angle , for eg: 90,45, -45.

lower_limit

default is NULL,assign the column if it exist in the dataframe otherwise omit

upper_limit

default is NULL,assign the column if it exist in the dataframe otherwise omit

hover_text

default NULL, give the info on values of x and y axis how fill if if needed.it could be designed as needed for eg:paste("%x|%Y-%m-%d","
count:%y")#'

legend_orientation

default = 'h' , change to 'v' if needed

legend_x

default = 0.5, changed when needed depending on the legendpostion.

legend_y

default = -0.35, changed when needed depending on the legendpostion.

show_legend

boolean, defaults to true,

Value

a plotly plot object

Examples

if (FALSE) {
line_chart_plotly(df= df, x = date, y = count,grouping_col = character)
}
if (FALSE) {
line_chart_plotly(df= df, x = date, y = count,grouping_col = character,
hover_text = hover_template(paste("%{x|%Y-%m-%d}","<br>count:%{y}<extra></extra>")),
y_label = "count/rate", x_label = "date",xaxis_label_angle  = -45/90/45,
yaxis_label_angle  = 0(default),x_axis_reverse= default(FALSE),y_rangemode = "tozero",
plot_mode = "lines"/"lines+markers",legend_orientation = 'h'/'v',
legend_x =0.5, legend_y = -0.35,show_legend = FALSE,
lower_limit = "lower_error", upper_limit = "upper_error")
# if df has lower and upper limit otherwise can omit it.
}


if (FALSE) {
line_chart_plotly(df= df, x = year(as.factor), y = rate,grouping_col = character,
hover_text = "%{x} year, %{y} rate<br>",
y_label = "count/rate", x_label = "date",xaxis_label_angle  = -45/90/45,
yaxis_label_angle  = 0(default),x_axis_reverse= default(FALSE),y_rangemode = "tozero",
plot_mode = "lines"/"lines+markers",legend_orientation = 'h'/'v',
legend_x =0.5, legend_y = -0.35,show_legend = FALSE,
lower_limit = "lower_error", upper_limit = "upper_error")
# if df has lower and upper limit #' otherwise can omit it.
}