epi_map

The map function has 2 mandatory arguments to plot, df (data frame containing your data) and value_col (value variable name), with additional arguments as described further below:

epi_map(
  df,
  value_col = NULL,
  inc_shp = TRUE,
  shp_filepath = NULL,
  shp_areacode = NULL,
  data_areacode = NULL,
  lookup = NULL,
  break_intervals = NULL,
  force_cat = TRUE,
  n_breaks = NULL,
  legend_title = NULL,
  legend_pos = "left",
  border_shape_filepath = NULL,
  border_code_col = NULL,
  border_areaname = NULL
)

Arguments

df

name of data frame for plotting

value_col

name of value variable in data frame

inc_shp

boolean parameter to indicate whether df includes shapefile data or if this will be uploaded separately.

shp_filepath

filepath for the shapefile required for spatial data in function. Note this will not be uploaded if inc_shp = TRUE.

shp_areacode

name of areacode variable in shapefile (Mandatory if shp_filepath argument passed).

data_areacode

name of areacode variable in dataframe (Mandatory if shp_filepath argument passed).

lookup

vector of categories to include in the legend.

break_intervals

numeric vector of interval points for the legend (Mandatory if lookup argument passed).

force_cat

boolean parameter to determine whether all arguments passed in the lookup argument are used, even if there are no values present in the data.

n_breaks

number of break intervals. This argument is an alternative to supplying defined breaks in lookup argument and will provide a number of evenly distributed breaks as specified. Note if lookup argument is passed this will be ignored.

legend_title

string containing the legend title.

legend_pos

string containing legend position.

border_shape_filepath

filepath for the shapefile if an outer border is required for the map. This should be a higher geography e.g. if creating an UTLA map- a regional shapefile can be used.

border_code_col

name of areacode variable in the border shapefile. Required if a specific area within the border shapefile is required.

border_areaname

string containing the name of area required for border shapefile (required if border_code_col argument passed).

Examples