Line_chart

The chart function has 3 mandatory arguments to plot, with additional arguments as described further below:

line_chart(
  df,
  base = NULL,
  x = NULL,
  y = NULL,
  group_var = NULL,
  line_colour = c("#003b49"),
  error_colour = c("#f2c75c"),
  ci = NULL,
  lower = NULL,
  upper = NULL,
  y_axis = "y1",
  h_line = NULL,
  line_type = "solid",
  add_points = NULL,
  line_size = 1,
  x_label = NULL,
  y_label = NULL,
  x_label_angle = NULL,
  y_label_angle = NULL,
  x_labels_reverse = NULL,
  y_min_limit = NULL,
  y_max_limit = NULL,
  x_axis_breaks = NULL,
  legend_pos = "bottom",
  remove_gridlines = NULL,
  percent = NULL,
  cap_text = NULL,
  no_shift = FALSE
)

Arguments

df

name of data frame for plotting

base

a ggplot instance or NULL if creating a new plot

x

name of x axis variable in data frame

y

name of y axis variable in data frame

group_var

name of group variable to generate multiple lines (if required) in data frame

line_colour

colour (or vector of colours if group_var != NULL) for the line in the chart

error_colour

if ci != NULL - colour (or vector of colours if group_var != NULL) for the error bar or ribbon in the chart

ci

"e" for error bars or "ribbon" for a ribbon - lower and upper become mandatory if ci != NULL

lower

lower value for error \ ribbon geom (mandatory if ci argument passed)

upper

upper value for error \ ribbon geom (mandatory if ci argument passed)

y_axis

y axis to use - y1 or y2 (default is y1 - standard left-hand axis, y2 can be used for dual-axis graphs)

h_line

will display a horizontal line if valid inter passed

line_type

if provided will change all lines used to that type i.e. dotted, dashed, default sitting is solid, when not provided where multiple lines generated each will be of different type

add_points

will simply add points to all lines generated

line_size

will alter line width accordingly for all lines generated, defaults to 1

x_label

for provision of an x axis label

y_label

for provision of an y axis label

x_label_angle

to adjust the x axis label by the degrees of the integer provided

y_label_angle

to adjust the y axis label by the degrees of the integer provided

x_labels_reverse

enter an argument of any value i.e. 'y' to reverse the x labeling order when using categorical data

y_min_limit

set the limit on the y axis scaling by proving an integer

y_max_limit

set the limit on the x axis scaling by proving an integer

x_axis_breaks

modify the x axis breaks by providing an integer

legend_pos

modify the position of the legend (where applicable) with appropriate value i.e. bottom (default position), top, right, left

remove_gridlines

enter an argument of any value i.e. 'y' to remove the grid lines

percent

enter an argument of any value i.e. 'y' to include the % symbol for y axis labels

cap_text

enter text for a caption to appear below plot

no_shift

If no shift should be applied to the secondary y-axis

Value

the final plot

Details

data frame x axis variable name y axis variable name

Examples